+++ front/lib/reinforcement/utils.ts
