+++ front/temporal/worker_registry.ts
@@ -14,7 +14,7 @@ import { runProductionChecksWorker } from "@app/temporal/production_checks/worke
 import { runProjectTodoWorker } from "@app/temporal/project_todo/worker";
 import { runUserProjectDigestQueueWorker } from "@app/temporal/project_user_digest_queue/worker";
 import { runReinforcedAgentWorker } from "@app/temporal/reinforced_agent/worker";
-import { runReinforcedSkillsWorker } from "@app/temporal/reinforced_skills/worker";
+import { runReinforcedSkillsWorker } from "@app/temporal/reinforcement/worker";
 import { runRelocationWorker } from "@app/temporal/relocation/worker";
 import { runRemoteToolsSyncWorker } from "@app/temporal/remote_tools/worker";
 import { runSandboxReaperWorker } from "@app/temporal/sandbox_reaper/worker";
@@ -44,7 +44,7 @@ export type WorkerName =
   | "poke"
   | "production_checks"
   | "reinforced_agent"
-  | "reinforced_skills"
+  | "reinforcement"
   | "project_user_digest_queue"
   | "relocation"
   | "sandbox_reaper"
@@ -71,7 +71,7 @@ export const workerFunctions: Record<WorkerName, () => Promise<void>> = {
   poke: runPokeWorker,
   production_checks: runProductionChecksWorker,
   reinforced_agent: runReinforcedAgentWorker,
-  reinforced_skills: runReinforcedSkillsWorker,
+  reinforcement: runReinforcedSkillsWorker,
   project_user_digest_queue: runUserProjectDigestQueueWorker,
   project_todo: runProjectTodoWorker,
   relocation: runRelocationWorker,
